deprecated_feature_notice: Move code to a separate module.
